+++ src/simulation/AntSimulation.java
@@ -2,7 +2,6 @@ package simulation;
 
 import java.util.ArrayList;
 import java.util.List;
-
 import cell.Ant;
 import cell.AntSpaceCell;
 import cell.Cell;
@@ -10,15 +9,16 @@ import cell.CellWithAngleAndPatch;
 import grid.GridOfCells;
 import javafx.util.Pair;
 
+
 /**
  * The Class AntSimulation sets the rules for the simulation regarding ants searching for food.
  */
-public class AntSimulation extends SimulationWithAngleAndPatch{
+public class AntSimulation extends SimulationWithAngleAndPatch {
     private int myCrowdedLevel;
     private int myPheromoneThreshold;
     private int myLeaveFood;
     private int myLeaveHome;
-    
+
     /**
      * Instantiates a new ant simulation.
      *
@@ -26,14 +26,16 @@ public class AntSimulation extends SimulationWithAngleAndPatch{
      * @param parameters the parameters
      */
     public AntSimulation (GridOfCells cellSocietyGrid, String[] parameters) {
-        super(cellSocietyGrid,Integer.MAX_VALUE);
-        myCrowdedLevel=Integer.parseInt(parameters[0]);
-        myPheromoneThreshold=Integer.parseInt(parameters[1]);
-        myLeaveFood=Integer.parseInt(parameters[2]);
-        myLeaveHome=Integer.parseInt(parameters[3]);
+        super(cellSocietyGrid, Integer.MAX_VALUE);
+        myCrowdedLevel = Integer.parseInt(parameters[0]);
+        myPheromoneThreshold = Integer.parseInt(parameters[1]);
+        myLeaveFood = Integer.parseInt(parameters[2]);
+        myLeaveHome = Integer.parseInt(parameters[3]);
     }
 
-    /* (non-Javadoc)
+    /*
+     * (non-Javadoc)
+     *
      * @see Simulation#processNeighbors(Cell, int, int)
      */
     @Override
@@ -41,7 +43,7 @@ public class AntSimulation extends SimulationWithAngleAndPatch{
         AntSpaceCell cell = (AntSpaceCell) currentCell;
         List<Cell> neighbors = getCellSocietyGrid().getNeighbors(x, y);
         patchMovement(currentCell, neighbors);
-        for (Ant a: cell.getCurrentAnts()) {
+        for (Ant a : cell.getCurrentAnts()) {
             if (a.hasFood()) {
                 goToNest(cell, a, x, y);
             }
@@ -60,18 +62,19 @@ public class AntSimulation extends SimulationWithAngleAndPatch{
      * @param y the y
      */
     private void goToNest (AntSpaceCell currentCell, Ant a, int x, int y) {
-        AntSpaceCell next = getMovement(currentCell, a, x, y, AntSpaceCell.HOMEINDEX, AntSpaceCell.FOOD);
-        if (next!=null) {
-            a.leaveFoodPheromone(currentCell, myPheromoneThreshold,myLeaveFood);
+        AntSpaceCell next =
+                getMovement(currentCell, a, x, y, AntSpaceCell.HOMEINDEX, AntSpaceCell.FOOD);
+        if (next != null) {
+            a.leaveFoodPheromone(currentCell, myPheromoneThreshold, myLeaveFood);
             a.setOrientation(next, currentCell.getXCoordinate(), currentCell.getYCoordinate());
             next.getFutureAnts().add(a);
             currentCell.getFutureAnts().remove(a);
-            if (next.getCurrentState()==AntSpaceCell.HOME) {
+            if (next.getCurrentState() == AntSpaceCell.HOME) {
                 a.setHasFood(false);
             }
         }
     }
-    
+
     /**
      * Go to food.
      *
@@ -81,18 +84,19 @@ public class AntSimulation extends SimulationWithAngleAndPatch{
      * @param y the y
      */
     private void goToFood (AntSpaceCell currentCell, Ant a, int x, int y) {
-        AntSpaceCell next = getMovement(currentCell, a, x, y, AntSpaceCell.FOODINDEX, AntSpaceCell.HOME);
-        if (next!=null) {
+        AntSpaceCell next =
+                getMovement(currentCell, a, x, y, AntSpaceCell.FOODINDEX, AntSpaceCell.HOME);
+        if (next != null) {
             a.leaveHomePheromone(currentCell, myPheromoneThreshold, myLeaveHome);
             a.setOrientation(next, currentCell.getXCoordinate(), currentCell.getYCoordinate());
             next.getFutureAnts().add(a);
             currentCell.getFutureAnts().remove(a);
-            if (next.getCurrentState()==AntSpaceCell.FOOD) {
+            if (next.getCurrentState() == AntSpaceCell.FOOD) {
                 a.setHasFood(true);
             }
         }
     }
-    
+
     /**
      * Gets the cell to move to
      *
@@ -104,45 +108,53 @@ public class AntSimulation extends SimulationWithAngleAndPatch{
      * @param antSpaceState the ant space state
      * @return the movement
      */
-    public AntSpaceCell getMovement(AntSpaceCell currentCell, Ant a, int x, int y, int patchIndex, int antSpaceState) {
+    public AntSpaceCell getMovement (AntSpaceCell currentCell,
+                                     Ant a,
+                                     int x,
+                                     int y,
+                                     int patchIndex,
+                                     int antSpaceState) {
         List<Cell> neighbors = getCellSocietyGrid().getNeighbors(x, y);
         List<Cell> forwardNeighbors = processNeighborAngle(neighbors, a, 90);
-        Cell maxHomePheromones = findMaxPatch(neighbors,null,patchIndex);
-        if (currentCell.getCurrentState()==antSpaceState) {
-            if (maxHomePheromones!=null) {
-                a.setOrientation(maxHomePheromones, 
+        Cell maxHomePheromones = findMaxPatch(neighbors, null, patchIndex);
+        if (currentCell.getCurrentState() == antSpaceState) {
+            if (maxHomePheromones != null) {
+                a.setOrientation(maxHomePheromones,
                                  currentCell.getXCoordinate(),
                                  currentCell.getYCoordinate());
             }
         }
-        AntSpaceCell next = (AntSpaceCell)findMaxPatch(forwardNeighbors,null,patchIndex);
-        if (next==null) {
-            next = (AntSpaceCell)maxHomePheromones;
+        AntSpaceCell next = (AntSpaceCell) findMaxPatch(forwardNeighbors, null, patchIndex);
+        if (next == null) {
+            next = (AntSpaceCell) maxHomePheromones;
         }
-        if (next==null) {
-            next = (AntSpaceCell)neighbors.get((int)(neighbors.size()*Math.random()));
+        if (next == null) {
+            next = (AntSpaceCell) neighbors.get((int) (neighbors.size() * Math.random()));
         }
         return next;
     }
-    
-    /* (non-Javadoc)
+
+    /*
+     * (non-Javadoc)
+     *
      * @see SimulationWithAngleAndPatch#processNeighborAngle(java.util.List, Cell, double)
      */
     @Override
     public List<Cell> processNeighborAngle (List<Cell> neighbors, Cell cCell, double angle) {
         CellWithAngleAndPatch cell = (CellWithAngleAndPatch) cCell;
         List<Cell> neighborsInRange = new ArrayList<Cell>();
-        List<Pair<Integer,Integer>> directions = new ArrayList<Pair<Integer,Integer>>();
-        for (int i=-1; i<=1; i++) {
-            int x = (int)(Math.round(Math.cos((cell.getAngle()+i*45)*Math.PI/180)));
-            int y = (int)(Math.round(Math.sin((cell.getAngle()+i*45)*Math.PI/180)));
-            directions.add(new Pair<Integer,Integer>(x,y));
+        List<Pair<Integer, Integer>> directions = new ArrayList<Pair<Integer, Integer>>();
+        for (int i = -1; i <= 1; i++) {
+            int x = (int) (Math.round(Math.cos((cell.getAngle() + i * 45) * Math.PI / 180)));
+            int y = (int) (Math.round(Math.sin((cell.getAngle() + i * 45) * Math.PI / 180)));
+            directions.add(new Pair<Integer, Integer>(x, y));
         }
-        for (Cell neighborCell: neighbors) {
+        for (Cell neighborCell : neighbors) {
             AntSpaceCell c = (AntSpaceCell) neighborCell;
-            if (c.getFutureAnts().size()<myCrowdedLevel) {
-                for (Pair<Integer,Integer> p: directions) {
-                    if (c.getXCoordinate()==(cCell.getXCoordinate()+(int)p.getKey()) && c.getYCoordinate()==(cCell.getYCoordinate()+(int)p.getValue())) {
+            if (c.getFutureAnts().size() < myCrowdedLevel) {
+                for (Pair<Integer, Integer> p : directions) {
+                    if (c.getXCoordinate() == (cCell.getXCoordinate() + p.getKey()) &&
+                        c.getYCoordinate() == (cCell.getYCoordinate() + p.getValue())) {
                         neighborsInRange.add(c);
                     }
                 }
@@ -150,8 +162,10 @@ public class AntSimulation extends SimulationWithAngleAndPatch{
         }
         return neighborsInRange;
     }
-    
-    /* (non-Javadoc)
+
+    /*
+     * (non-Javadoc)
+     *
      * @see Simulation#updateCurrentStates()
      */
     @Override
